Update css to work better with Firefox
The widget css was a bit off with firefox (compared to chrome).
Problem was the line height and the default font that was set.

This is fixed by specifying a font family (seems that both arial &
verdana are cross-platform) and use e relative line height instead of a
rem unit.