Fix gulp not minifying the css
Since Gulp doesn't wait for one task to be completed in order to execute
the second one, we hade some troubles with our default gulp task: the
css wasn't minified because there was no css to minify in the dist
folder yet.
This commit fixes this by adding the "css" task to the "minify-css"
task. So "css" is executed first and then the "minify-css" task is
executed if you call the "minify-css" task.